requirements.txt
notebook==7.2.2
torch==2.5.1
numpy<2
transformers==4.46.3
datasets==3.1.0
accelerate==1.1.1
python-dotenv==1.0.1
protobuf==5.29.0
pandas==2.2.3
tuned-lens==0.2.0
pre-commit==4.0.1
matplotlib==3.9.3
circuitsvis==1.41.0
loguru==0.7.3
rouge-score==0.1.2
typer[all]==0.9.0
h5py==3.12.1
python-decouple==3.8
wandb==0.18.7
scikit-learn==1.6.1
PyQt5==5.15.10
Pygments==2.18.0
huggingface-hub==0.26.3
fasttext==0.9.3
anywidget==0.9.18
kaleido==0.2.1
